Clean-up tests; increase code coverage
